parrillada-gio-media: Eject empty drive with no error message
When an empty drive is ejected the user is presented an error dialog when
the I/O opeartion times out - this is due to the fact that the
GDrive::changed signal is not emitted when an empty drive is ejected and
therefore brasero_gio_operation_end() isn't called.

Work around this by checking if the drive is empty in
brasero_gio_operation_eject_finish () and ending the I/O opeartion if it is.

https://bugzilla.gnome.org/show_bug.cgi?id=701730

Backport from https://git.gnome.org/browse/brasero/commit/?id=f3d0b989b41ec0cee4d551f1ca02687db98a333d